Fix precision < number length
I didn't check for the length of the number when I set width when
precision was specified, which led to numbers being truncated if the
precision was also less than the number length.

Also fixed a small bug where if the precision and the width were the
same, and they were both less than the number, it wouldn't update the
width at all (to the correct value of the number length).