Set precision to len if it is less
Making this change fixed both the issues from the previous commit, so I
undid the previous changes. Now precision will always be set to at least
length, which means if width is the same (i.e. it passes the >=), then
it doesn't need set (so I can remove the = part). If it is less than
precision, I only need to set it to precision, since precision is at
least the size of len (so I remove the call to ft_max).

I also remove the string, since I 1) ran out of lines and 2) can simply
return the function call to ft_strinit. I didn't need the variable.