Change struct to contain the conversion character
The struct has been changed to contain the conversion character. The
index is still returned from get_conversion, for use in the jump table.
the dispatch function has also been edited to take this parameter. This
makes it easier to see the conversion character in later functions,
which will be useful with bBoOxX