Fix problems in the decimal part
I adjusted the if statement that determines the length of the fraction
to consider a number that will be rounded up to 1 at the end, which
cleans up a bit at the end of that function. It also resolves an issue
with my previous "fix" that added an extra number to the decimal part in
some cases.