Handle precision and width with unicode strings
I have added my wide string to utf-8 function in this flag handler so I
can change it to use precision (didn't want to make a whole new function
for it in the libft, and I also didn't want to have to deal with getting
the length for that nonsense after it already went through the function.
I also updated flag_wstr and flag_wchar to deal with width, with a
function called addspaces, which adds spaces on the left or right side
depending on a parameter.